Function pallet_treasury::pallet::dispatchables::payout

pub fn payout<T: Config<I>, I: 'static>(index: SpendIndex)

Claim a spend.

Dispatch Origin

Must be signed

Details

Spends must be claimed within some temporal bounds. A spend may be claimed within one Config::PayoutPeriod from the valid_from block. In case of a payout failure, the spend status must be updated with the check_status dispatchable before retrying with the current function.

Parameters

index: The spend index.

Events

Emits Event::Paid if successful.

Warning: Doc-Only

This function is an automatically generated, and is doc-only, uncallable stub. See the real version in Pallet::payout.