Function pallet_treasury::pallet::dispatchables::void_spend

pub fn void_spend<T: Config<I>, I: 'static>(index: SpendIndex)

Void previously approved spend.

Dispatch Origin

Must be Config::RejectOrigin.

Details

A spend void is only possible if the payout has not been attempted yet.

Parameters

index: The spend index.

Events

Emits Event::AssetSpendVoided if successful.

Warning: Doc-Only

This function is an automatically generated, and is doc-only, uncallable stub. See the real version in Pallet::void_spend.