Type Alias pallet_treasury::pallet::Module

pub type Module<T, I = ()> = Pallet<T, I>;

👎Deprecated: use Pallet instead

Type alias to Pallet, to be used by construct_runtime.

Generated by pallet attribute macro.

Aliased Type

struct Module<T, I = ()>(/* private fields */);

Implementations

impl<T: Config<I>, I: 'static> Pallet<T, I>

pub fn propose_spend( origin: OriginFor<T>, value: BalanceOf<T, I>, beneficiary: <<T as Config>::Lookup as StaticLookup>::Source ) -> DispatchResult

👎Deprecated: propose_spend will be removed in February 2024. Use spend instead.

Put forward a suggestion for spending.

Dispatch Origin

Must be signed.

Details

A deposit proportional to the value is reserved and slashed if the proposal is rejected. It is returned once the proposal is awarded.

Complexity

O(1)

Events

Emits Event::Proposed if successful.

pub fn reject_proposal( origin: OriginFor<T>, proposal_id: ProposalIndex ) -> DispatchResult

👎Deprecated: reject_proposal will be removed in February 2024. Use spend instead.

Reject a proposed spend.

Dispatch Origin

Must be Config::RejectOrigin.

Details

The original deposit will be slashed.

Complexity

O(1)

Events

Emits Event::Rejected if successful.

pub fn approve_proposal( origin: OriginFor<T>, proposal_id: ProposalIndex ) -> DispatchResult

👎Deprecated: approve_proposal will be removed in February 2024. Use spend instead.

Approve a proposal.

Dispatch Origin

Must be Config::ApproveOrigin.

Details

At a later time, the proposal will be allocated to the beneficiary and the original deposit will be returned.

Complexity

O(1).

Events

No events are emitted from this dispatch.

pub fn spend_local( origin: OriginFor<T>, amount: BalanceOf<T, I>, beneficiary: <<T as Config>::Lookup as StaticLookup>::Source ) -> DispatchResult

Propose and approve a spend of treasury funds.

Dispatch Origin

Must be Config::SpendOrigin with the Success value being at least amount.

Details

NOTE: For record-keeping purposes, the proposer is deemed to be equivalent to the beneficiary.

Parameters

amount: The amount to be transferred from the treasury to the beneficiary.

beneficiary: The destination account for the transfer.

Events

Emits Event::SpendApproved if successful.

pub fn remove_approval( origin: OriginFor<T>, proposal_id: ProposalIndex ) -> DispatchResult

Force a previously approved proposal to be removed from the approval queue.

Dispatch Origin

Must be Config::RejectOrigin.

Details

The original deposit will no longer be returned.

Parameters

proposal_id: The index of a proposal

Complexity

O(A) where A is the number of approvals

Errors

Error::ProposalNotApproved: The proposal_id supplied was not found in the approval queue, i.e., the proposal has not been approved. This could also mean the proposal does not exist altogether, thus there is no way it would have been approved in the first place.

pub fn spend( origin: OriginFor<T>, asset_kind: Box<T::AssetKind>, amount: AssetBalanceOf<T, I>, beneficiary: Box<<<T as Config<I>>::BeneficiaryLookup as StaticLookup>::Source>, valid_from: Option<BlockNumberFor<T>> ) -> DispatchResult

Propose and approve a spend of treasury funds.

Dispatch Origin

Must be Config::SpendOrigin with the Success value being at least amount of asset_kind in the native asset. The amount of asset_kind is converted for assertion using the Config::BalanceConverter.

Details

Create an approved spend for transferring a specific amount of asset_kind to a designated beneficiary. The spend must be claimed using the payout dispatchable within the Config::PayoutPeriod.

Parameters

asset_kind: An indicator of the specific asset class to be spent.

amount: The amount to be transferred from the treasury to the beneficiary.

beneficiary: The beneficiary of the spend.

valid_from: The block number from which the spend can be claimed. It can refer to the past if the resulting spend has not yet expired according to the Config::PayoutPeriod. If None, the spend can be claimed immediately after approval.

Events

Emits Event::AssetSpendApproved if successful.

pub fn payout(origin: OriginFor<T>, index: SpendIndex) -> DispatchResult

Claim a spend.

Dispatch Origin

Must be signed.

Details

Spends must be claimed within some temporal bounds. A spend may be claimed within one Config::PayoutPeriod from the valid_from block. In case of a payout failure, the spend status must be updated with the check_status dispatchable before retrying with the current function.

Parameters

index: The spend index.

Events

Emits Event::Paid if successful.

pub fn check_status( origin: OriginFor<T>, index: SpendIndex ) -> DispatchResultWithPostInfo

Check the status of the spend and remove it from the storage if processed.

Dispatch Origin

Must be signed.

Details

The status check is a prerequisite for retrying a failed payout. If a spend has either succeeded or expired, it is removed from the storage by this function. In such instances, transaction fees are refunded.

Parameters

index: The spend index.

Events

Emits Event::PaymentFailed if the spend payout has failed. Emits Event::SpendProcessed if the spend payout has succeed.

pub fn void_spend(origin: OriginFor<T>, index: SpendIndex) -> DispatchResult

Void previously approved spend.

Dispatch Origin

Must be Config::RejectOrigin.

Details

A spend void is only possible if the payout has not been attempted yet.

Parameters

index: The spend index.

Events

Emits Event::AssetSpendVoided if successful.

impl<T: Config<I>, I: 'static> Pallet<T, I>

pub fn proposal_count() -> ProposalIndex

An auto-generated getter for ProposalCount.

impl<T: Config<I>, I: 'static> Pallet<T, I>

pub fn proposals<KArg>( k: KArg ) -> Option<Proposal<T::AccountId, BalanceOf<T, I>>>where KArg: EncodeLike<ProposalIndex>,

An auto-generated getter for Proposals.

impl<T: Config<I>, I: 'static> Pallet<T, I>

pub fn approvals() -> BoundedVec<ProposalIndex, T::MaxApprovals>

An auto-generated getter for Approvals.

impl<T: Config<I>, I: 'static> Pallet<T, I>

pub fn account_id() -> T::AccountId

The account ID of the treasury pot.

This actually does computation. If you need to keep using it, then make sure you cache the value and only call this once.

pub fn spend_funds() -> Weight

Spend some money! returns number of approvals before spend.

pub fn pot() -> BalanceOf<T, I>

Return the amount of money in the pot.