Trait pallet_treasury::Config

pub trait Config<I = DefaultInstance>: Config {
    type PalletId: Get<PalletId>;
    type Currency: Currency<Self::AccountId> + ReservableCurrency<Self::AccountId>;
    type ApproveOrigin: EnsureOrigin<Self::Origin>;
    type RejectOrigin: EnsureOrigin<Self::Origin>;
    type Event: From<Event<Self, I>> + Into<Self::Event>;
    type OnSlash: OnUnbalanced<NegativeImbalanceOf<Self, I>>;
    type ProposalBond: Get<Permill>;
    type ProposalBondMinimum: Get<BalanceOf<Self, I>>;
    type SpendPeriod: Get<Self::BlockNumber>;
    type Burn: Get<Permill>;
    type BurnDestination: OnUnbalanced<NegativeImbalanceOf<Self, I>>;
    type WeightInfo: WeightInfo;
    type SpendFunds: SpendFunds<Self, I>;
    type MaxApprovals: Get<u32>;
}

Associated Types

type PalletId: Get<PalletId>

The treasury’s module id, used for deriving its sovereign account ID.

type Currency: Currency<Self::AccountId> + ReservableCurrency<Self::AccountId>

The staking balance.

type ApproveOrigin: EnsureOrigin<Self::Origin>

Origin from which approvals must come.

type RejectOrigin: EnsureOrigin<Self::Origin>

Origin from which rejections must come.

type Event: From<Event<Self, I>> + Into<Self::Event>

The overarching event type.

type OnSlash: OnUnbalanced<NegativeImbalanceOf<Self, I>>

Handler for the unbalanced decrease when slashing for a rejected proposal or bounty.

type ProposalBond: Get<Permill>

Fraction of a proposal’s value that should be bonded in order to place the proposal. An accepted proposal gets these back. A rejected proposal does not.

type ProposalBondMinimum: Get<BalanceOf<Self, I>>

Minimum amount of funds that should be placed in a deposit for making a proposal.

type SpendPeriod: Get<Self::BlockNumber>

Period between successive spends.

type Burn: Get<Permill>

Percentage of spare funds (if any) that are burnt per spend period.

type BurnDestination: OnUnbalanced<NegativeImbalanceOf<Self, I>>

Handler for the unbalanced decrease when treasury funds are burned.

type WeightInfo: WeightInfo

Weight information for extrinsics in this pallet.

type SpendFunds: SpendFunds<Self, I>

Runtime hooks to external pallet using treasury to compute spend funds.

type MaxApprovals: Get<u32>

The maximum number of approvals that can wait in the spending queue.

Loading content...