Trait pallet_treasury::Trait

pub trait Trait<I = DefaultInstance>: Trait {
    type ModuleId: Get<ModuleId>;
    type Currency: Currency<Self::AccountId> + ReservableCurrency<Self::AccountId>;
    type ApproveOrigin: EnsureOrigin<Self::Origin>;
    type RejectOrigin: EnsureOrigin<Self::Origin>;
    type Tippers: Contains<Self::AccountId> + ContainsLengthBound;
    type TipCountdown: Get<Self::BlockNumber>;
    type TipFindersFee: Get<Percent>;
    type TipReportDepositBase: Get<<Self::Currency as Currency<Self::AccountId>>::Balance>;
    type DataDepositPerByte: Get<<Self::Currency as Currency<Self::AccountId>>::Balance>;
    type Event: From<Event<Self, I>> + Into<Self::Event>;
    type OnSlash: OnUnbalanced<<Self::Currency as Currency<Self::AccountId>>::NegativeImbalance>;
    type ProposalBond: Get<Permill>;
    type ProposalBondMinimum: Get<<Self::Currency as Currency<Self::AccountId>>::Balance>;
    type SpendPeriod: Get<Self::BlockNumber>;
    type Burn: Get<Permill>;
    type BountyDepositBase: Get<<Self::Currency as Currency<Self::AccountId>>::Balance>;
    type BountyDepositPayoutDelay: Get<Self::BlockNumber>;
    type BountyUpdatePeriod: Get<Self::BlockNumber>;
    type BountyCuratorDeposit: Get<Permill>;
    type BountyValueMinimum: Get<<Self::Currency as Currency<Self::AccountId>>::Balance>;
    type MaximumReasonLength: Get<u32>;
    type BurnDestination: OnUnbalanced<<Self::Currency as Currency<Self::AccountId>>::NegativeImbalance>;
    type WeightInfo: WeightInfo;
}

Associated Types

type ModuleId: Get<ModuleId>

The treasury's module id, used for deriving its sovereign account ID.

type Currency: Currency<Self::AccountId> + ReservableCurrency<Self::AccountId>

The staking balance.

type ApproveOrigin: EnsureOrigin<Self::Origin>

Origin from which approvals must come.

type RejectOrigin: EnsureOrigin<Self::Origin>

Origin from which rejections must come.

type Tippers: Contains<Self::AccountId> + ContainsLengthBound

Origin from which tippers must come.

ContainsLengthBound::max_len must be cost free (i.e. no storage read or heavy operation).

type TipCountdown: Get<Self::BlockNumber>

The period for which a tip remains open after is has achieved threshold tippers.

type TipFindersFee: Get<Percent>

The percent of the final tip which goes to the original reporter of the tip.

type TipReportDepositBase: Get<<Self::Currency as Currency<Self::AccountId>>::Balance>

The amount held on deposit for placing a tip report.

type DataDepositPerByte: Get<<Self::Currency as Currency<Self::AccountId>>::Balance>

The amount held on deposit per byte within the tip report reason or bounty description.

type Event: From<Event<Self, I>> + Into<Self::Event>

The overarching event type.

type OnSlash: OnUnbalanced<<Self::Currency as Currency<Self::AccountId>>::NegativeImbalance>

Handler for the unbalanced decrease when slashing for a rejected proposal or bounty.

type ProposalBond: Get<Permill>

Fraction of a proposal's value that should be bonded in order to place the proposal. An accepted proposal gets these back. A rejected proposal does not.

type ProposalBondMinimum: Get<<Self::Currency as Currency<Self::AccountId>>::Balance>

Minimum amount of funds that should be placed in a deposit for making a proposal.

type SpendPeriod: Get<Self::BlockNumber>

Period between successive spends.

type Burn: Get<Permill>

Percentage of spare funds (if any) that are burnt per spend period.

type BountyDepositBase: Get<<Self::Currency as Currency<Self::AccountId>>::Balance>

The amount held on deposit for placing a bounty proposal.

type BountyDepositPayoutDelay: Get<Self::BlockNumber>

The delay period for which a bounty beneficiary need to wait before claim the payout.

type BountyUpdatePeriod: Get<Self::BlockNumber>

Bounty duration in blocks.

type BountyCuratorDeposit: Get<Permill>

Percentage of the curator fee that will be reserved upfront as deposit for bounty curator.

type BountyValueMinimum: Get<<Self::Currency as Currency<Self::AccountId>>::Balance>

Minimum value for a bounty.

type MaximumReasonLength: Get<u32>

Maximum acceptable reason length.

type BurnDestination: OnUnbalanced<<Self::Currency as Currency<Self::AccountId>>::NegativeImbalance>

Handler for the unbalanced decrease when treasury funds are burned.

type WeightInfo: WeightInfo

Weight information for extrinsics in this pallet.

Loading content...