Crate pallet_treasury

Treasury Pallet

The Treasury pallet provides a “pot” of funds that can be managed by stakeholders in the system and a structure for making spending proposals from this pot.

Config

Call

Overview

The Treasury Pallet itself provides the pot to store funds, and a means for stakeholders to propose, approve, and deny expenditures. The chain will need to provide a method (e.g. inflation, fees) for collecting funds.

By way of example, the Council could vote to fund the Treasury with a portion of the block reward and use the funds to pay developers.

Terminology

Proposal: A suggestion to allocate funds from the pot to a beneficiary.

Beneficiary: An account who will receive the funds from a proposal iff the proposal is approved.

Deposit: Funds that a proposer must lock when making a proposal. The deposit will be returned or slashed if the proposal is approved or rejected respectively.

Pot: Unspent funds accumulated by the treasury pallet.

Interface

Dispatchable Functions

General spending/proposal protocol:

propose_spend - Make a spending proposal and stake the required deposit.

reject_proposal - Reject a proposal, slashing the deposit.

approve_proposal - Accept the proposal, returning the deposit.

remove_approval - Remove an approval, the deposit will no longer be returned.

GenesisConfig

The Treasury pallet depends on the GenesisConfig.

Re-exports

pub use weights::WeightInfo;

pub use pallet::*;

Modules

pallet

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.

weights

Autogenerated weights for pallet_treasury

Structs

Proposal

A spending proposal.

Traits

SpendFunds

A trait to allow the Treasury Pallet to spend it’s funds for other purposes. There is an expectation that the implementer of this trait will correctly manage the mutable variables passed to it:

Type Definitions

BalanceOf

NegativeImbalanceOf

PositiveImbalanceOf

ProposalIndex

An index of a proposal. Just a u32.