Crate pallet_treasury

Treasury Module

The Treasury module provides a "pot" of funds that can be managed by stakeholders in the system and a structure for making spending proposals from this pot.

treasury::Trait

Call

Overview

The Treasury Module itself provides the pot to store funds, and a means for stakeholders to propose, approve, and deny expenditures. The chain will need to provide a method (e.g. inflation, fees) for collecting funds.

By way of example, the Council could vote to fund the Treasury with a portion of the block reward and use the funds to pay developers.

Terminology

Proposal: A suggestion to allocate funds from the pot to a beneficiary.

Beneficiary: An account who will receive the funds from a proposal iff the proposal is approved.

Deposit: Funds that a proposer must lock when making a proposal. The deposit will be returned or slashed if the proposal is approved or rejected respectively.

Pot: Unspent funds accumulated by the treasury module.

Interface

Dispatchable Functions

propose_spend - Make a spending proposal and stake the required deposit.

set_pot - Set the spendable balance of funds.

configure - Configure the module's proposal requirements.

reject_proposal - Reject a proposal, slashing the deposit.

approve_proposal - Accept the proposal, returning the deposit.

GenesisConfig

The Treasury module depends on the GenesisConfig.

Structs

GenesisConfig

Module

Proposal
A spending proposal.

Enums

Call
Dispatchable calls.

Error
Error for the treasury module.

RawEvent
Events for this module.

Traits

Trait

Type Definitions

Event
RawEvent specialized for the configuration Trait