pallet_treasury::pallet

Trait Config

pub trait Config<I: 'static = ()>: Config + Config<RuntimeEvent: From<Event<Self, I>>> {

    type Currency: Currency<Self::AccountId> + ReservableCurrency<Self::AccountId>;
    type ApproveOrigin: EnsureOrigin<Self::RuntimeOrigin>;
    type RejectOrigin: EnsureOrigin<Self::RuntimeOrigin>;
    type RuntimeEvent: From<Event<Self, I>> + IsType<<Self as Config>::RuntimeEvent>;
    type OnSlash: OnUnbalanced<NegativeImbalanceOf<Self, I>>;
    type ProposalBond: Get<Permill>;
    type ProposalBondMinimum: Get<BalanceOf<Self, I>>;
    type ProposalBondMaximum: Get<Option<BalanceOf<Self, I>>>;
    type SpendPeriod: Get<BlockNumberFor<Self>>;
    type Burn: Get<Permill>;
    type PalletId: Get<PalletId>;
    type BurnDestination: OnUnbalanced<NegativeImbalanceOf<Self, I>>;
    type WeightInfo: WeightInfo;
    type SpendFunds: SpendFunds<Self, I>;
    type MaxApprovals: Get<u32>;

}

Configuration trait of this pallet.

The main purpose of this trait is to act as an interface between this pallet and the runtime in which it is embedded in. A type, function, or constant in this trait is essentially left to be configured by the runtime that includes this pallet.

Consequently, a runtime that wants to include this pallet must implement this trait.

Required Associated Types

type Currency: Currency<Self::AccountId> + ReservableCurrency<Self::AccountId>

The staking balance.

type ApproveOrigin: EnsureOrigin<Self::RuntimeOrigin>

Origin from which approvals must come.

type RejectOrigin: EnsureOrigin<Self::RuntimeOrigin>

Origin from which rejections must come.

type RuntimeEvent: From<Event<Self, I>> + IsType<<Self as Config>::RuntimeEvent>

The overarching event type.

type OnSlash: OnUnbalanced<NegativeImbalanceOf<Self, I>>

Handler for the unbalanced decrease when slashing for a rejected proposal or bounty.

type ProposalBond: Get<Permill>

Fraction of a proposal’s value that should be bonded in order to place the proposal. An accepted proposal gets these back. A rejected proposal does not.

type ProposalBondMinimum: Get<BalanceOf<Self, I>>

Minimum amount of funds that should be placed in a deposit for making a proposal.

type ProposalBondMaximum: Get<Option<BalanceOf<Self, I>>>

Maximum amount of funds that should be placed in a deposit for making a proposal.

type SpendPeriod: Get<BlockNumberFor<Self>>

Period between successive spends.

type Burn: Get<Permill>

Percentage of spare funds (if any) that are burnt per spend period.

type PalletId: Get<PalletId>

The treasury’s pallet id, used for deriving its sovereign account ID.

type BurnDestination: OnUnbalanced<NegativeImbalanceOf<Self, I>>

Handler for the unbalanced decrease when treasury funds are burned.

type WeightInfo: WeightInfo

Weight information for extrinsics in this pallet.

type SpendFunds: SpendFunds<Self, I>

Runtime hooks to external pallet using treasury to compute spend funds.

type MaxApprovals: Get<u32>

The maximum number of approvals that can wait in the spending queue.

NOTE: This parameter is also used within the Bounties Pallet extension if enabled.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.