Trait pallet_treasury::Trait

pub trait Trait: Trait {
    type Currency: Currency<Self::AccountId> + ReservableCurrency<Self::AccountId>;
    type ApproveOrigin: EnsureOrigin<Self::Origin>;
    type RejectOrigin: EnsureOrigin<Self::Origin>;
    type Event: From<Event<Self>> + Into<Self::Event>;
    type ProposalRejection: OnUnbalanced<<Self::Currency as Currency<Self::AccountId>>::NegativeImbalance>;
    type ProposalBond: Get<Permill>;
    type ProposalBondMinimum: Get<<Self::Currency as Currency<Self::AccountId>>::Balance>;
    type SpendPeriod: Get<Self::BlockNumber>;
    type Burn: Get<Permill>;
}

Associated Types

type Currency: Currency<Self::AccountId> + ReservableCurrency<Self::AccountId>

The staking balance.

type ApproveOrigin: EnsureOrigin<Self::Origin>

Origin from which approvals must come.

type RejectOrigin: EnsureOrigin<Self::Origin>

Origin from which rejections must come.

type Event: From<Event<Self>> + Into<Self::Event>

The overarching event type.

type ProposalRejection: OnUnbalanced<<Self::Currency as Currency<Self::AccountId>>::NegativeImbalance>

Handler for the unbalanced decrease when slashing for a rejected proposal.

type ProposalBond: Get<Permill>

Fraction of a proposal's value that should be bonded in order to place the proposal. An accepted proposal gets these back. A rejected proposal does not.

type ProposalBondMinimum: Get<<Self::Currency as Currency<Self::AccountId>>::Balance>

Minimum amount of funds that should be placed in a deposit for making a proposal.

type SpendPeriod: Get<Self::BlockNumber>

Period between successive spends.

type Burn: Get<Permill>

Percentage of spare funds (if any) that are burnt per spend period.

Loading content...